moved getLists function to own file
